requirements.txt
Gradio
# Transformers
llama-cpp-python
huggingface_hub
python-dotenv
ddgs
yfinance
langchain_groq

